% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{daily2cal}
\alias{daily2cal}
\title{Create Calendar from Daily File}
\usage{
daily2cal(
  path = "daily.txt",
  start = NULL,
  end = NULL,
  days = NULL,
  delim = "::\\\\s+"
)
}
\arguments{
\item{path}{path to "daily" file}

\item{start}{start date}

\item{end}{end date}

\item{days}{a character string or a numeric vector indicating
which days of the week to use by default.
(1 = N = Sunday, 5 = R = Thursday, 7 = S = Saturday)}

\item{delim}{delimeter for parsing compents of a day's entry}
}
\value{
a data frame representing the calendar
}
\description{
Create Calendar from Daily File
}
\seealso{
\code{\link[=html_calendar]{html_calendar()}}, \code{\link[=gg_calendar]{gg_calendar()}}
}
