% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{daily2html}
\alias{daily2html}
\title{Render Daily File as HTML}
\usage{
daily2html(path = "daily.txt", start = NULL, end = NULL,
  days = NULL, show = 1:7, items = NULL, width = "95\%")
}
\arguments{
\item{path}{path to "daily" file}

\item{start}{start date}

\item{end}{end date}

\item{days}{a character string or a numeric vector indicating
which days of the week to use by default.
(1 = N = Sunday, 5 = R = Thursday, 7 = S = Saturday)}

\item{show}{a numeric vector of days to show (Sunday = 1)}

\item{items}{a character vector of items to display (or NULL for all)}

\item{width}{width of HTML table used for calendar.}
}
\value{
a character string of HTML code for the calendar
}
\description{
Render Daily File as HTML
}
