% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{gg_calendar}
\alias{gg_calendar}
\title{Create gg Calendar}
\usage{
gg_calendar(calendar, show = 1:7, size = 3)
}
\arguments{
\item{calendar}{a calendar object}

\item{show}{a numeric vector of days to show (Sunday = 1)}

\item{size}{a number for scaling text size}
}
\value{
a ggplot object
}
\description{
Create gg calendar
}
