% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{html_calendar}
\alias{html_calendar}
\title{Create HTML Calendar}
\usage{
html_calendar(calendar, show = 1:7, items = NULL, width = "95\%")
}
\arguments{
\item{calendar}{a calendar object}

\item{show}{a numeric vector of days to show (Sunday = 1)}

\item{items}{a character vector of items to display (or NULL for all)}

\item{width}{width of HTML table used for calendar.}
}
\value{
a character string of HTML code for the calendar
}
\description{
Create HTML Calendar
}
