% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{daily2cal}
\alias{daily2cal}
\title{Create Calendar from Daily File}
\usage{
daily2cal(path = "daily.txt", start = lubridate::today() -
  lubridate::weeks(2), end = lubridate::today() + lubridate::weeks(4),
  mask = 2:6)
}
\arguments{
\item{path}{path to "daily" file}

\item{start}{start date}

\item{end}{end date}

\item{mask}{a numeric vector indicating which days of the week to include. (1 = Sunday)}
}
\value{
a data frame representing the calendar
}
\description{
Create Calendar from Daily File
}
\seealso{
\code{\link[=html_calendar]{html_calendar()}}, \code{\link[=gg_calendar]{gg_calendar()}}
}
